pyproject.toml
[project]
name = "trading-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair<5",
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=1.27.0",
    "numpy<2",
    "optuna>=4.7.0",
    "pandas<2.2.3",
    "protobuf<5.0.0",
    "pyarrow<15",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "setuptools<70.0.0",
    "streamlit>=1.19.0",
    "yfinance>=1.2.0",
]
